improve how we run the tests
